fix(viewer): split the BCF browser-OAuth flow out of bcf-server.ts; plain punctuation in sign-in copy
bcf-server.ts crossed the 400-line module ratchet with the redirect-URI
override, so the browser OAuth half (prepare/complete, redirect URI,
preparation type) moves to bcf-server-oauth.ts and the shared sign-in
tail plus the lazy bcf-api loader to bcf-server-session.ts; bcf-server.ts
re-exports both, so no caller or test changes.

User-facing strings in the connect form and the sign-in errors drop em
dashes and curly apostrophes for plain sentences.

The corrupt-storage test now silences the console.warn that echoed the
parser's SyntaxError into the TAP stream, which the changed-tests-observe
gate read as a module load failure and refused to judge the branch on.
